feat: Update type defs for query transformer
