+++ packages/cubejs-api-gateway/package.json
@@ -15,6 +15,7 @@
     "test": "jest"
   },
   "main": "index.js",
+  "types": "index.d.ts",
   "dependencies": {
     "@hapi/joi": "^15.1.1",
     "chrono-node": "^1.3.11",
