+++ packages/cubejs-server-core/core/index.d.ts
@@ -3,6 +3,7 @@ import {
   Response as ExpressResponse,
   NextFunction as ExpressNextFunction
 } from "express";
+import { NormalizedQuery as Query } from "@cubejs-backend/api-gateway";
 
 declare module "@cubejs-backend/server-core" {
   export function create(options?: CreateOptions): any;
@@ -99,51 +100,4 @@ declare module "@cubejs-backend/server-core" {
     | "redshift"
     | "snowflake"
     | "sqlite";
-
-  export interface QueryFilter {
-    member: string;
-    operator:
-      | "equals"
-      | "notEquals"
-      | "contains"
-      | "notContains"
-      | "gt"
-      | "gte"
-      | "lt"
-      | "lte"
-      | "set"
-      | "notSet"
-      | "inDateRange"
-      | "notInDateRange"
-      | "beforeDate"
-      | "afterDate";
-    values?: string[];
-  }
-
-  export type QueryTimeDimensionGranularity =
-    | "hour"
-    | "day"
-    | "week"
-    | "month"
-    | "year";
-
-  export interface QueryTimeDimension {
-    dimension: string;
-    dateRange?: string[] | string;
-    granularity?: QueryTimeDimensionGranularity;
-  }
-
-  export interface Query {
-    measures: string[];
-    dimensions?: string[];
-    filters?: QueryFilter[];
-    timeDimensions?: QueryTimeDimension[];
-    segments?: string[];
-    limit?: number;
-    offset?: number;
-    order?: "asc" | "desc";
-    timezone?: string;
-    renewQuery?: boolean;
-    ungrouped?: boolean;
-  }
 }
